Remove godeps
Commit 4719b674 switched the repository to use the dep tool for vendor
management, but it didn't remove the Godeps directory. Commit a61d332
then updated the golang.org/x/net dependency using godep, but not dep.

This commit updates the locked dep config so that it picks the same
version as godep, then removes godep so there is only one tool for
managing the repository. It's generated using dep v0.5.0, which changes
the fields used in the lockfile. Note that the contents of vendor/ do
not change in any significant way in this commit.